Rename default packages to get rid of underscores
