--- src/main/java/io/github/i_vrnv/github_scoring/GithubScoringApplication.java
@@ -1,13 +1,0 @@
-package io.github.i_vrnv.github_scoring;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GithubScoringApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GithubScoringApplication.class, args);
-	}
-
-}
