+++ src/main/java/io/github/ivrnv/github/scoring/GithubScoringApplication.java
@@ -1,0 +1,13 @@
+package io.github.ivrnv.github.scoring;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GithubScoringApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GithubScoringApplication.class, args);
+	}
+
+}
