--- src/test/java/io/github/i_vrnv/github_scoring/GithubScoringApplicationTests.java
@@ -1,13 +1,0 @@
-package io.github.i_vrnv.github_scoring;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GithubScoringApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
