+++ src/test/java/io/github/ivrnv/github/scoring/GithubScoringApplicationTests.java
@@ -1,0 +1,13 @@
+package io.github.ivrnv.github.scoring;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GithubScoringApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
